requirements.txt
asgiref==3.7.2
boto3==1.26.165
botocore==1.29.165
dj-database-url==2.0.0
Django==4.2.2
django-storages==1.13.2
jmespath==1.0.1
Pillow==10.0.0
psycopg2==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.0
tzdata==2023.3
urllib3==1.26.16
